Ensure a unique cache key for searches
The cache key is the JSON-stringified version of `searches`.
For `glean metrics` the cache keys for triaged and untriaged searches
were the same, because filtering is done client-side, not in the search.
Toggling between tabs thus lead to data being read from the cache for
both searches.

Adding a _unique_ (but unused) entry in the search helps to
differentiate between them.