Actually cache responses to bug API queries
Due to parallelism, each group of queries hitting `this.findBugs` could fall
into `await findBugs` while the cache remained empty.

Instead, let's cache the Promise-ed data. Then all of the groups waiting on the
same data will be waiting on a single request. Takes 13 BMO queries down to 9,
and 6 github calls down to 4.

(( Browser-level caching may actually have helped a smidge here, but also maybe
not. Depended on the server config, I think. ))

Also uncovered a caching bug. The cache key didn't account for filtering logic
(because JSON.stringify will not stringify JS functions). The requests were for
the same data, but cached the filtered data. "Luckily", each of these were
getting their own copies of the unfiltered data, so it didn't matter...

...until the cache started working properly, of course.

So now we cache the API result instead of the filtered result.

Still plenty more that could be improved here, but this is as much time as I
want to spend here.